Extensive changes: add threading to the server, introduce generic "pseudo-packets"
When the server accepts a socket connection, it spins off a thread to
handle that connection.  Threads assure cleanup of file-handles, and
the server assures cleanup of threads.

"pseudo-packets" on the socket stream potentially allow out-of-band
comms.  (e.g. to support the old MarFS "abort" concept, or so client
can optionally request an incremental checksum, without always
requiring the server to compute that).  Not sure if this is a great
idea: if followed to the limit, all communication would require an
"op", followed by a size_t (e.g. length), followed by optional data
(e.g. data having the indicated length).  For now, once we've
communicated a GET/PUT op, using length to refer to the length of the
filename, the data itself does not get "packetized" like this.