Added RDMA via riowrite().
The read-side has no way of knowing when the write has completed, so I
introduced a simple synchronizing protocol.  The writer sends a "DATA"
pseudo-packet after an RDMA transfer, indicating the size of the data,
and the reader sends an "ACK" when the buffer is no longer needed,
allowing the next write to begin.

In the case of a PUT from libne to an MC block-file, both sides will
have work to do between the synchronization events, so this shouldn't
add much latency to individual threads.  However, this scheme can be
extended to support double-buffering, if we decide we need it.