Alternative build to integrate the RDMA-sockets client/server with libne.
This requires that libne now do some of the path-formatting internally
to the functions that deal with paths, such as ne_open().  This is
because variations across block-numbers in libne can now imply
variations in the host and directory-tree for each block-file.  Solved
by passing in a function-pointer (which performs the sprintf to
produce the final path), and a corresponding pointer to some arbitrary
state, which can be used by the function.  MarFS callers can supply a
state that includes configuration details, and a function-pointer
which knows how to use those details to compute pathnames.  Thus,
libne can generate the full host-and-path for each block-file.  The
default libne implementations just use a default snprintf function, to
provide the previous functionality.