major re-factor allows code re-use between client-side and server-side.
The client-side of a GET and server-side of a PUT (and vice versa) are
essentially the same (except client also sends "GET/PUT" commands).
Reusing code assures that the protocols and RDMA initializations are
done the same in both directions.

With RDMA, the writer must not write more than the buffer size of the
reader, so the GET/PUT protocols now have reader send buff-size to
writer.  libne expects application buff-size to be moved in one shot,
so we interpose a routine to repeatedly call skt_write/skt_read, until
the entire buffer has been sent/received.

open/close and clean-up handlers can now be used by both client and
server.