Allow changing default client/server buf-sizes at config-time.  Fix config bugs.
Server-side bufs, used for transfers between sockets and server-side
files, are currently allocated on the stack, to simplify thread
clean-up.  Use "S_BUFSIZE=..."  on the config command-line, to change
the default.

Almost all client-side bufs (and their sizes) are specified as
arguments in function-calls, so there is no need to configure those.
The exceptions are client_get()/client_put(), which are currently just
demo functions in client.c.  Use "C_BUFSIZE=..." to change.