Added a server-side reaper-thread to clean-up stalled connections.
Killing a pftool job kills all the client-side socket connections to
the server, but it seems that the system allows the server-side
threads to remain stuck in read/write, instead of getting
disconnected.

So, every so often (currently 10 sec), the reaper checks all existing
server-side connections for data-movement, since the previous check.
Threads managing connections that haven't moved data since the last
check are cancelled.

TBD: The server is sometimes leaking file-descriptors.  Not sure yet
whether the reaper-thread is responsible (but it seems plausible).
Added some tests to 'client' to help chase this.