Interim merge of rdma_auth, which shows that threads resolve an RDMA+fuse issue
The setresuid in the push/pop of user-identities at entry/exit from
fuse routines (in MarFS fuse) interferes with the ability of a thread
to make new connections over infiniband, including IB+RDMA using
rsockets.  The "threads" branch of libne (if built with
--enable-sockets=rdma) keeps SocketHandles local to custom-created
threads that are doing the I/O.  This avoids having the push/pop be
done in the same thread(s) in which the connections are maintained,
which seems to work.

Because the merge involves some conflicts, I'm checking in this
interim merge of rdma_auth, with conflicts resolved, to avoid having
to redo that work with future merges.