client/server S3-authentication working for single threads.
A fixed user-name (currently "root") is looked up in ~/.awsAuth on
client and server.  The "password" found there is used by the client
to sign a blob comprised of user-name, date, operation, and filename,
which (signature) is included along with the comprising fields, in the
request to the server.  The server then does its own lookup, assures
that the date is not too old, generates its own signature, and
compares with the client's signature.  A match authenticates that
operation.

TBD:

-- There is a skt_fcntl() which would allow credentials obtained by a
client with escalated priviledges (e.g. when fuse is initializing the
DAL), to be installed later into freshly opened socket-handles, after
the client has de-escalated.  Not tested yet.  This should be exported
through something like "ne_fcntl()", to allow the MC-sockets DAL to
authenticate libne transactions (e.g. ne_open, ne_fcntl, ne_write,
..., ne_close).

-- allow timezone differences between client and server