propagate S3 authentication for RDMA-sockets into libne proper
The issue for libne is that the aws4c authentication-info used by the
RDMA-socket interface under ./sockets/ must be passed in at run-time,
because it requires privileges to access, and (in the case of
marfs/pftool) the client-side runs de-escalated.  So, this info must
be cached at start-up by the client, and passed in later.

I'm not completely happy with what this does to the libne interface.
I've preserved the old interface, for apps that have the option of
running as root, so they can just look up auth-info at runtime,
instead of breaking.

We depend on a recent update to libwas4c (42cb1ac), which now
exports GetStringToSign(), needed to generate S3 hashes used in
authentication.

Configure with --disable-auth, if you aren't using sockets and don't
want to have to link with libaws4c.