Respect authentication by users other than SKT_S3_AUTH.
Where this is headed is e.g. allowing Jenkins to have its own
~/.awsAuth, which would allow authenticated access only to regions of
the data store that were accessible by that user.  This exploits
previously-unused support for per-repo user-names in the MC_SOCKETS DAL.

The test_client now tries to authenticate first as $USER, looking in
~/.awsAuth.  For now, the server runs as root, and simply validates
against its local /root/.awsAuth, so any user authentication is
sufficient to access anything.  But the next step would be to have the
server seteuid before authentication, so an authenticated access could
only access files for that repo.