Adding experimental fast_timer library (libft) using TSC for low-overhead.
test_fast_timer.c is built as an example app.  This suggests that the
cost of fast_timer_start() + fast_timer_stop() is ~70 ns.  (The reason
it isn't faster is that we track the current CPU and core, so we can
ignore intervals during which a thread migrates across TSCs.)

Among other things, I'm hoping this will allow us to unobtrustively
maintain some stats on the dynamic behavior of N+E parallel channels
in libne writes, to identify slowpokes, network performance issues,
etc.