Fix copy_file_to_socket() to avoid wasted file-reads in the event of a seek.
Reorganized copy_file_to_socket() to look for seeks in the
control-stream before doing file-reads.  This crops up in the case of
a client doing a seek on a GET stream from the server.  libne does
this fairly frequently.

This branch takes a "peek" approach (as apposed to an "unread"
approach), which turns out to be much cleaner.