Situation after weeks of chasing a race condition.
The race condition turned out to lie within librdmacm.  We have a
(simple) patch pending.  This code works against that patched library.
Without the fix, pftool will occasionally segfault.

There is commented-out locking (see LOCK/UNLOCK in skt_common.c) which
may eventually allow us to achieve a similar fix against the unpatched
current-release (librdmacm-1.0.21), while we're waiting for the patch
to make it into a RHEL update but that is not working yet.