Added PRINTlog(), so libneTest output can get printed in non-debug builds.
You'd think PRINTerr() in libne internals should just go to stderr,
but in non-debug builds, doing that leads to pftool spewing alarming
low-level errors to users, in the case where e.g. libne is timing-out
and retrying, etc.  Thus, we suppress the "stderr" output for
non-debug builds.  PRINTlog() is the thing to use, when you really-o
truely-o want your output to got to the user, even the pftool user.