Added a (primitive) server-side test API.  Requires building with --enable-test-api
This is not a "back door".  Interaction with the test-API still goes
through the usual authentication, and test functions do not
(currently) provide special access to data, or change the behavior of
the server.  However, there's no need for this API for normal
operation, so unless you build with --enable-test-api it will not be
available.