Make fast_timer_inits() thread-safe.  Add fast_timer_msec().
Thread-safety is normally not an issue in apps that can call that
once, up front.  The locking for thread-safety is never touched,
except by the first thread or two that ever call the function.  This
change supports pending changes to read_raw()/write_raw(), where they
might want to use a fast timer for cheap reconing of total elapsed
time.