Recent tests show rpoll() does see messages sent before rpoll() was entered.
This would seem to allow us to simplify read_raw()/write_raw().
However, rpoll() does not detect the situation when the peer
shuts-down unexpectedly (e.g. someone kill pftool).  In that case, we
wait out the full RD/WR_TIMEOUT, before punting.  If these leaves
server threads out of action for too long, we may want to break up the
timeout-interval into a series of short polls, maybe trying the
read/write (non-blocking) in between, as was done in the older
implementation.