High-throughput testing with RDMA to SSD store showed server-side also vulnerable to librdma-cm bug
This is the bug we've patched in the rmda-core code, but on RHEL6
systems it can affect the server.  This commit wraps locking around
the raccept(), just like we did elsewhere with open() and close().
This has some awkward implications for dynamics, as current locking is
just a mutex.  Therefore, the server can be unresponsive while an
exiting thread holds the lock for ~30 sec, in rclose() -> rshutdown()
-> read().  This can happen if e.g. pftool gets cancelled.  Working
on improvements, but these changes are a start.