Build generates sockets/skt_config.h, instead of setting up compile-time '-D' args
The issue is that sockets/skt_common.h defines structs that vary,
depending on #defines that are set up in the configuration.  If the
config just guides the formation of the Makefile to produce these
defines at compile-time, then external systems (e.g. MarFS) which also
include this file would have to assure that *their* config/make
produces the same flags.

A much safer approach is for us to generate a top-level include-file,
which contains the configured #defines.  Then external systems
reliably get our configured options at their compile-time.