generate config-info .h file even for non-sockets builds.
sockets/skt_config.h is now config/udal_config.h, and it is generated
and moved to the install for all builds.  You'd think it would've been
obvious that it shouldn't live in sockets/, if we want this file to be
built with non-sockets builds, as well.  The stupid mistake was hidden
by versions of the file residual from previous sockets builds.