move va_end() up out of parse_va_open_args() into callers.
va_end() is actually a macro that is supposed to be in the function
that calls va_start().  Lei and I thought we saw some crashes from
missing va_end(), in the past, and so I made these changes while
chasing down a crash that was actually fixed by the preceeding commit.