DAL: add new flags arg to verify functions (#30)
This commit changes the DAL verify functions to take a flag argument
instead of a single char value. The flags will allow passing multiple
boolean parameters to the verify function.

In addition to the pre-existing fix paramter, this adds a new
check owner parameter that determines whether the verify function should
check the UID/GID of the secure root dir. This parameter will be false
when starting MarFS, because MarFS can run under a different user than
the directory owner, but will be true when verifying the config.